requirements.txt
aiogram>=3.4.1
python-dotenv>=1.0.1

SQLAlchemy>=2.0.30
aiosqlite>=0.20.0

aiohttp>=3.9.5

matplotlib>=3.8.4
pydantic>=2.7.1

# опционально: перевод RU->EN (если включишь TRANSLATE_ENABLED=1)
deep-translator>=1.11.4
